Do step for delete options
